Allow exception-less pipeline failure

+++ Refresher.Core/Pipelines/Steps/OutputElfStep.cs
@@ -1,4 +1,6 @@
-﻿namespace Refresher.Core.Pipelines.Steps;
+﻿using Refresher.Core.Platform;
+
+namespace Refresher.Core.Pipelines.Steps;
 
 public class OutputElfStep : Step
 {
@@ -14,8 +16,10 @@ public class OutputElfStep : Step
     public override Task ExecuteAsync(CancellationToken cancellationToken = default)
     {
         string elfOutput = this.Inputs.First().GetValueFromPipeline(this.Pipeline);
-        // if (File.Exists(elfOutput))
-            // TODO: ask user if they want to replace
+        if (File.Exists(elfOutput) && this.Platform.Ask("The output EBOOT already exists. Would you like to replace it?") == QuestionResult.No)
+        {
+            return Task.CompletedTask;
+        }
 
         File.Copy(this.Game.DecryptedEbootPath!, Path.GetFullPath(elfOutput));
